+++ puissance4.h
@@ -3,12 +3,14 @@
 #define PLAYER1 1
 #define PLAYER2 2
 
-extern int grid[NB_LINE][NB_COLUMN];
+#include <stdint.h>
+
+extern uint8_t grid[NB_LINE][NB_COLUMN];
 
 void display_grid(void);
-int play(int column, int player);
-int win_horizontal(int line, int player);
-int win_vertical(int column, int player);
-int win_diagonal(int line, int column, int player);
-int check_win(int line, int column, int player);
-int is_full(int column);
+uint8_t play(uint8_t column, uint8_t player);
+uint8_t win_horizontal(uint8_t line, uint8_t player);
+uint8_t win_vertical(uint8_t column, uint8_t player);
+uint8_t win_diagonal(uint8_t line, uint8_t column, uint8_t player);
+uint8_t check_win(uint8_t line, uint8_t column, uint8_t player);
+uint8_t is_full(uint8_t column);
